feat: added base login and register
