+++ frontend/src/app/auth/page.tsx
@@ -1,82 +1,14 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 "use client";
 
-import type React from "react";
-
-import { useState } from "react";
-import Link from "next/link";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Bookmark } from "lucide-react";
-import { authClient } from "@/lib/auth-client";
-import { useRouter } from "next/navigation";
+import { Card, CardHeader } from "@/components/ui/card";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import LoginForm from "./components/login-form";
+import RegistrationForm from "./components/registration-form";
+import { useAuth } from "./services/auth-service";
 
 export default function AuthPage() {
-  const [isLoading, setIsLoading] = useState(false);
-  const router = useRouter();
-
-  const handleLogin = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const email = "asda@asdsad.com";
-    const password = "asdasdasd";
-    const { data, error } = await authClient.signIn.email(
-      {
-        email,
-        password,
-      },
-      {
-        onRequest: (ctx) => {
-          setIsLoading(true);
-        },
-        onSuccess: (ctx) => {
-          setIsLoading(false);
-          router.push("/dashboard");
-        },
-        onError: (ctx) => {
-          alert(ctx.error.message);
-          console.log(ctx);
-        },
-      },
-    );
-  };
-
-  const handleRegister = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const email = "admin@example.com";
-    const password = "password123";
-    const name = "Admin User";
-
-    const { data, error } = await authClient.signUp.email(
-      {
-        email,
-        password,
-        name,
-      },
-      {
-        onRequest: (ctx) => {
-          setIsLoading(true);
-        },
-        onSuccess: (ctx) => {
-          setIsLoading(false);
-          router.push("/dashboard");
-        },
-        onError: (ctx) => {
-          alert(ctx.error.message);
-          console.log(ctx);
-        },
-      },
-    );
-  };
+  const { isLoading, handleLogin, handleRegister } = useAuth();
 
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-50 px-4">
@@ -98,86 +30,14 @@ export default function AuthPage() {
             </CardHeader>
 
             <TabsContent value="login">
-              <form onSubmit={handleLogin}>
-                <CardContent className="space-y-4">
-                  <CardTitle>Welcome back</CardTitle>
-                  <CardDescription>
-                    Enter your credentials to access your bookmarks
-                  </CardDescription>
-                  <div className="space-y-2">
-                    <Label htmlFor="email">Email</Label>
-                    <Input
-                      id="email"
-                      type="email"
-                      placeholder="Enter your email"
-                      required
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="password">Password</Label>
-                    <Input
-                      id="password"
-                      type="password"
-                      placeholder="Enter your password"
-                      required
-                    />
-                  </div>
-                </CardContent>
-                <CardFooter className="flex flex-col space-y-4">
-                  <Button type="submit" className="w-full" disabled={isLoading}>
-                    {isLoading ? "Signing in..." : "Sign in"}
-                  </Button>
-                  <Link
-                    href="#"
-                    className="text-sm text-blue-600 hover:underline"
-                  >
-                    Forgot your password?
-                  </Link>
-                </CardFooter>
-              </form>
+              <LoginForm handleLogin={handleLogin} isLoading={isLoading} />
             </TabsContent>
 
             <TabsContent value="register">
-              <form onSubmit={handleRegister}>
-                <CardContent className="space-y-4">
-                  <CardTitle>Create account</CardTitle>
-                  <CardDescription>
-                    Sign up to start organizing your bookmarks
-                  </CardDescription>
-                  <div className="space-y-2">
-                    <Label htmlFor="name">Full Name</Label>
-                    <Input
-                      id="name"
-                      type="text"
-                      placeholder="Enter your full name"
-                      required
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="email-register">Email</Label>
-                    <Input
-                      id="email-register"
-                      type="email"
-                      placeholder="Enter your email"
-                      required
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="password-register">Password</Label>
-                    <Input
-                      id="password-register"
-                      type="password"
-                      placeholder="Create a password"
-                      required
-                    />
-                  </div>
-                </CardContent>
-                <CardFooter>
-                  <Button type="submit" className="w-full" disabled={isLoading}>
-                    {isLoading ? "Creating account..." : "Create account"}
-                  </Button>
-                </CardFooter>
-              </form>
+              <RegistrationForm
+                handleRegister={handleRegister}
+                isLoading={isLoading}
+              />
             </TabsContent>
           </Tabs>
         </Card>
